evaluation/requirements.txt
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-openai>=0.3.0,<0.4.0

openai>=1.0.0,<3.0.0

transformers>=4.37.0
accelerate>=0.25.0
sentencepiece>=0.1.99

faiss-cpu>=1.7.4

SQLAlchemy>=1.4
python-dotenv>=1.0.0
requests>=2.31.0
nltk>=3.8

django>=4.2,<6.0